fix(dialog): modify MDCDialog subcomponents to always return a VNode
Modify MDCDialog subcomponents to return an empty VNode instead of
`undefined` when nothing is in their default slot. While returning
`undefined` from a render function does not result in a runtime
error, it does cause the linter to throw an error, since the
render function's signature does not match the signature defined
in Vue.