fix(ripple): set root element CSS classes using `cssClass` data property
Set the `mdc-ripple` component's root element's CSS classes using the
`cssClass` data property.

This change addresses an issue that would arise when the consumer of the
`mdc-ripple` component sets CSS classes on top of the classes added by
the component internally, and later modifies the class list using a
Vue-provided mechanism. When this happens, the `mdc-ripple-upgraded`
class on the root element, added by the `MDCRippleFoundation` via the
`addClass` adapter method, is overwritten. This is because the
`addClass` method modifies the root element's class list using
`classList.add()`, completely circumventing Vue, and allowing any
class added via the `addClass` method to be ignored when the component
renders again.