fix(multiple): remove attrs with nullish values from linkable components
Remove attributes with nullish values from linkable components.

Linkable components can be provided with the tag of a custom Vue
component to render as their root element via the `to` prop. This is
useful for components like `RouterLink` and `NuxtLink`. If the provided
tag does not equal `'a'`, then the `href` computed property is `null`.
If this `null` value is assigned to the `href` attribute of the element
the custom Vue component is mounted to, then this value will override
the value that the custom Vue component may be setting it to.
Effectively, this results in the `href` attribute being blank,
regardless of what the custom Vue component is trying to set it as.